Sending Beacons

mPulseAPI.jl uses the mPulse REST Beacon API to send beacons.

You first need to make a config request using mPulseAPI.getBeaconConfig and then include that config in your call to mPulseAPI.sendBeacon along with other beacon parameters.

Beacon parameters may be named by label or name.

mPulseAPI.CONFIG_URL — Constant

Base URL for config.json requests

mPulseAPI.getBeaconConfig — Method

Fetch beacon configuration for an mPulse APP using the Beacon API Caches the result for whatever is specified in Cache-control: max-age

mPulseAPI.sendBeacon — Method

Send a beacon to mPulse