Object caching

This module caches the results of GET requests to the mPulse REST API to improve performance. You can manipulate this cache using the following methods.

mPulseAPI.clearAlertCache — Method

Expire an entry from the alert cache. Use this if the alert has changed.

Keyword Arguments

alertID::Int64 : The ID of the alert to expire.

alertName::AbstractString : The Alert name in mPulse. This can be found from the mPulse alert configuration dialog.

Returns

true : on success

false : if the entry was not in cache

mPulseAPI.clearDomainCache — Method

Expire an entry from the domain cache. Use this if the domain has changed.

Keyword Arguments

domainID::Int64 : The ID of the domain to expire.

appKey::AbstractString : The App Key (formerly known as API key) associated with the domain. This is available from the mPulse domain configuration dialog.

appName::AbstractString : The App name in mPulse. This can be got from the mPulse domain configuration dialog.

Returns

true : on success

false : if the entry was not in cache

mPulseAPI.clearStatModelCache — Method

Expire an entry from the statistical model cache. Use this if the model has changed.

Keyword Arguments

statModelID::Int64 : The ID of the statistical model to expire.

statModelName::AbstractString : The statistical model name in mPulse. This can be found from the mPulse statistical model configuration dialog.

Returns

true : on success

false : if the entry was not in cache

mPulseAPI.clearTenantCache — Method

Expire an entry from the tenant cache. Use this if the tenant has changed.

Keyword Arguments

tenantID::Int64 : The ID of the tenant to expire.

name::AbstractString : The Tenant name in mPulse. This is got from the mPulse domain configuration dialog.

Returns

true : on success

false : if the entry was not in cache

mPulseAPI.clearTokenCache — Method

Expire an entry from the token cache. Use this if the token associated with this tenant is no longer valid.

Arguments

tenant::AbstractString : The tenant name whose token needs to be expired

Returns

true : on success

false : if the entry was not in cache