mPulseAPI.jl Documentation

Getting Started

mPulseAPI — Module

Communicate with the mPulse Query & Repository REST APIs to fetch information about tenants and apps.

Documentation

This module:

mPulseAPI.jl: https://akamai.github.io/mPulseAPI.jl/

REST API that this module uses:

mPulse API: https://techdocs.akamai.com/mpulse/reference/api

Quick and dirty usage

This snippet will get you up and running, see the full documentation for more details.

See how to generate an API Token for details about the apiToken

using mPulseAPI

# mPulse uses apiToken for authentication
token = getRepositoryToken("<tenant name>", "<mPulse api token for tenant>")

# Get a domain by app name
domain = getRepositoryDomain(token, appName="<app name from mPulse>")

# Get a domain by App Key (formerly known as API key)
domain = getRepositoryDomain(token, appKey="<App Key from mPulse>")

domain["attributes"]["appKey"]                           # Gets the App Key (formerly known as API key)
                                                         # for this app
domain["custom_metrics"]                                 # Get a Dict of custom metrics
domain["custom_metrics"]["Conversion Rate"]              # Get mapping for Conversion Rate custom metric
domain["custom_metrics"]["Conversion Rate"]["fieldname"] # Get field name for Conversion Rate custom
                                                         # metric

# Get all domains in tenant
domains = getRepositoryDomain(token)

# Get a tenant
tenant = getRepositoryTenant(token, name="<tenant name from mPulse>")

mPulseAPI.setVerbose — Function

Set verbosity of API calls.

If set to true, all URLs, headers and POST data will be printed to the console before making an API call.

Next steps

The following pages provide more detail on this module:

RepositoryAPI

Alerts & Anomaly Detection Models

QueryAPI

Additional APIs

Object caching

Exceptions

Index

mPulseAPI

mPulseAPI.mPulseAPIAuthException

mPulseAPI.mPulseAPIBugException

mPulseAPI.mPulseAPIException

mPulseAPI.mPulseAPIRequestException

mPulseAPI.mPulseAPIResultFormatException

mPulseAPI.clearAlertCache

mPulseAPI.clearAlertCache

mPulseAPI.clearDomainCache

mPulseAPI.clearDomainCache

mPulseAPI.clearStatModelCache

mPulseAPI.clearStatModelCache

mPulseAPI.clearTenantCache

mPulseAPI.clearTenantCache

mPulseAPI.clearTokenCache

mPulseAPI.deleteRepositoryAlert

mPulseAPI.deleteRepositoryObject

mPulseAPI.deleteRepositoryStatModel

mPulseAPI.getABTestTimers

mPulseAPI.getAPIResults

mPulseAPI.getBrowserTimers

mPulseAPI.getCustomMetricMap

mPulseAPI.getCustomTimerMap

mPulseAPI.getGeoTimers

mPulseAPI.getHistogram

mPulseAPI.getMetricOverPageLoadTime

mPulseAPI.getMetricsByDimension

mPulseAPI.getNodeContent

mPulseAPI.getPageGroupTimers

mPulseAPI.getRepositoryAlert

mPulseAPI.getRepositoryDomain

mPulseAPI.getRepositoryObject

mPulseAPI.getRepositoryStatModel

mPulseAPI.getRepositoryTenant

mPulseAPI.getRepositoryToken

mPulseAPI.getSessionsOverPageLoadTime

mPulseAPI.getSummaryTimers

mPulseAPI.getTimerByMinute

mPulseAPI.getTimersMetrics

mPulseAPI.mergeMetrics

mPulseAPI.postRepositoryAlert

mPulseAPI.postRepositoryObject

mPulseAPI.postRepositoryStatModel

mPulseAPI.setEndpoints

mPulseAPI.setVerbose